+++ examples/configure.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Install updates:
+apk -U upgrade --no-cache && rm -rf /var/cache/apk/*